refactor(network): abstract peer-announce emitter into own module
