+++ backend/src/protocol/network.ts
@@ -26,6 +26,7 @@ import { classifyConnection as classifyConnectionFn, dialProtocol as dialProtoco
 import { LISHServingHandlers, type SearchLishsMessage } from './lish-handlers.ts';
 export type { SearchLishsMessage } from './lish-handlers.ts';
 export { isSearchAdvertisableLish } from './lish-handlers.ts';
+import { PeerAnnounceManager, type PeerAnnounceMessage } from './peer-announce.ts';
 type PubSub = any; // PubSub type - using any since the exact type isn't exported from @libp2p/interface v3
 
 /**
@@ -55,46 +56,6 @@ interface PubsubEvent {
 	/** Cryptographically-verified peer ID of the original publisher (provided by libp2p gossipsub). */
 	from?: { toString(): string };
 }
-/**
- * Gossip-based peer-discovery bootstrap.
- *
- * Periodically each node broadcasts its reachable multiaddrs (and a subset of its
- * known peerStore) on every lishnet topic it is subscribed to. Receivers parse the
- * list and pass it through `addBootstrapPeers`, which dedupes against known peers
- * and calls `dial()`. This augments gossipsub PX (which only propagates on PRUNE)
- * and libp2p autodial (which is gated by peerStore) in topologies where bootstrap
- * hubs are few and NATed fleet members rely on relay reservations that expire
- * before libp2p would normally re-dial.
- */
-interface PeerAnnounceMessage {
-	type: 'peer-announce';
-	/** Multiaddrs (as strings) we claim to be reachable on, including /p2p/<peerID>. */
-	multiaddrs: string[];
-}
-/**
- * Adaptive peer-announce interval. Instead of a fixed cadence that spams the network
- * once saturated, the emitter picks an interval based on peerStore size — aggressive
- * when isolated, lazy when near full visibility. Traffic at saturation is reduced
- * roughly 6× compared to a fixed 20s cadence.
- */
-const PEER_ANNOUNCE_INTERVAL_ISOLATED_MS = 15_000; // peerStore < 20 (cold start / edge peer)
-const PEER_ANNOUNCE_INTERVAL_STEADY_MS = 30_000; // peerStore 20..80 (mid-convergence)
-const PEER_ANNOUNCE_INTERVAL_SATURATED_MS = 180_000; // peerStore > 80 (near full visibility)
-const PEER_ANNOUNCE_JITTER_RATIO = 0.25; // ±25% jitter of chosen interval (thunder-herd avoidance)
-/** Minimum peerStore size before we consider ourselves worth advertising. */
-const PEER_ANNOUNCE_MIN_PEER_STORE = 5;
-/** Hard cap on number of multiaddrs we include in a single announce (safety bound). */
-const PEER_ANNOUNCE_MAX_ADDRS = 32;
-/**
- * Cap on TOTAL multiaddrs in a peer-announce (self + peerStore transitive).
- * 128 covers ~half a 100-peer fleet per announce; receivers fill in the rest
- * from their own peerStore + subsequent announce cycles. Halving from 256
- * cuts saturation-time announce bandwidth ~50% with negligible discovery
- * latency cost (sub-2 announce cycles to fill peerStore).
- */
-const PEER_ANNOUNCE_MAX_TOTAL_ADDRS = 128;
-/** Max addrs we take from a single known peer when including transitive list. */
-const PEER_ANNOUNCE_MAX_ADDRS_PER_PEER = 3;
 /**
  * Handler for parsed pubsub topic messages.
  * `from` is the original publisher peer ID (verified by libp2p) when available —
@@ -131,7 +92,6 @@ export class Network {
 	private readonly dataServer: DataServer;
 	private readonly dataDir: string;
 	private statusInterval: NodeJS.Timeout | null = null;
-	private peerAnnounceInterval: NodeJS.Timeout | null = null;
 	/** Monotonic counter for status-interval ticks. Used by the periodic autodial promotion. */
 	private statusTickCount = 0;
 	/**
@@ -204,6 +164,9 @@ export class Network {
 	/** Handles incoming LISH-serving pubsub messages (want, searchLishs). */
 	private readonly lishHandlers: LISHServingHandlers;
 
+	/** Manages periodic peer-announce gossip emission and inbound peer-announce handling. */
+	private readonly peerAnnounce: PeerAnnounceManager;
+
 	/**
 	 * Per-peer re-dial backoff tracker. Re-dial attempts that fail bump the
 	 * per-peer failCount and push nextAttempt forward exponentially (30s × 2^fails
@@ -230,6 +193,12 @@ export class Network {
 			wantResponseCooldownMs: WANT_RESPONSE_COOLDOWN_MS,
 			getNode: () => this.node,
 			dialByPeerId: (peerID, protocol) => this.dialProtocolByPeerId(peerID, protocol),
+		});
+		this.peerAnnounce = new PeerAnnounceManager({
+			getNode: () => this.node,
+			getPubsub: () => this.pubsub,
+			broadcast: (topic, msg) => this.broadcast(topic, msg),
+			addBootstrapPeers: (multiaddrs, networkID, origin) => this.addBootstrapPeers(multiaddrs, networkID, origin),
 		});
 	}
 
@@ -479,7 +448,7 @@ export class Network {
 		this.setupBootstrapWorkaround();
 		this.setupStatusInterval();
 		this.setupWantResponseCleanup();
-		this.setupPeerAnnounceEmitter();
+		this.peerAnnounce.start();
 	}
 
 	// =========================================================================
@@ -653,174 +622,6 @@ export class Network {
 			}
 			if (searchRemoved > 0) trace(`[NET] search-dedup cleanup: pruned ${searchRemoved}, kept ${this.seenSearchIDs.size}`);
 		}, WANT_RESPONSE_CLEANUP_INTERVAL_MS);
-	}
-
-	/**
-	 * Periodically broadcast our reachable multiaddrs on every subscribed lishnet topic so
-	 * far-away peers who never encountered us via autodial/PX can still learn about us.
-	 * Gated on peerStore size so a freshly-started isolated node does not flood empty topics.
-	 * See `PeerAnnounceMessage` doc for the broader rationale.
-	 */
-	private setupPeerAnnounceEmitter(): void {
-		const schedule = async () => {
-			if (!this.node || !this.pubsub) return;
-			// Pick base interval from current peerStore saturation.
-			let base: number;
-			try {
-				const storeSize = (await this.node.peerStore.all()).length;
-				if (storeSize < 20) base = PEER_ANNOUNCE_INTERVAL_ISOLATED_MS;
-				else if (storeSize < 80) base = PEER_ANNOUNCE_INTERVAL_STEADY_MS;
-				else base = PEER_ANNOUNCE_INTERVAL_SATURATED_MS;
-			} catch {
-				base = PEER_ANNOUNCE_INTERVAL_STEADY_MS;
-			}
-			const jitter = Math.floor((Math.random() * 2 - 1) * base * PEER_ANNOUNCE_JITTER_RATIO);
-			const delay = Math.max(5_000, base + jitter);
-			this.peerAnnounceInterval = setTimeout(async () => {
-				try {
-					await this.emitPeerAnnounce();
-				} catch (err: any) {
-					trace(`[NET] peer-announce emit error: ${err?.message ?? err}`);
-				}
-				schedule().catch(() => {
-					/* schedule is async but errors handled inline */
-				});
-			}, delay);
-		};
-		schedule().catch(() => {
-			/* first-tick scheduling failure would leave emitter stopped — acceptable fallback */
-		});
-	}
-
-	private async emitPeerAnnounce(): Promise<void> {
-		if (!this.node || !this.pubsub) return;
-		const allPeers = await this.node.peerStore.all();
-		if (allPeers.length < PEER_ANNOUNCE_MIN_PEER_STORE) return;
-		// Include our full known peerStore in addition to our own reachable multiaddrs.
-		// This turns peer-announce into a transitive gossip protocol: edge-of-mesh
-		// peers learn about the whole fleet in one hop instead of waiting for mesh
-		// GRAFT to surface them. Total payload bounded by PEER_ANNOUNCE_MAX_TOTAL_ADDRS.
-		const collected = new Set<string>();
-		const localCidrs = getLocalCidrs();
-		let skippedSelf = 0;
-		let skippedTransitive = 0;
-		// Our own multiaddrs first (priority). Filter loopback (127.0.0.0/8) and
-		// non-local private addresses through shouldDenyDial — a remote peer
-		// receiving our /ip4/127.0.0.1 would otherwise TCP-loop to itself and
-		// hit identity-mismatch on every dial (validated on docker 2026-05-24:
-		// the moment debug logging landed it captured 3× back-to-back loopback
-		// dials from peer-announce intake within 3s of startup).
-		for (const ma of this.node.getMultiaddrs()) {
-			const s = ma.toString();
-			if (s.includes('/p2p-circuit')) continue;
-			if (shouldDenyDial(ma, localCidrs)) {
-				skippedSelf++;
-				continue;
-			}
-			collected.add(s);
-			if (collected.size >= PEER_ANNOUNCE_MAX_ADDRS) break;
-		}
-		// Transitive peerStore addrs.
-		const myID = this.node.peerId.toString();
-		for (const peer of allPeers) {
-			if (collected.size >= PEER_ANNOUNCE_MAX_TOTAL_ADDRS) break;
-			const pid = peer.id.toString();
-			if (pid === myID) continue;
-			let perPeer = 0;
-			for (const addr of peer.addresses) {
-				if (perPeer >= PEER_ANNOUNCE_MAX_ADDRS_PER_PEER) break;
-				if (collected.size >= PEER_ANNOUNCE_MAX_TOTAL_ADDRS) break;
-				const base = addr.multiaddr.toString();
-				if (base.includes('/p2p-circuit')) continue;
-				if (shouldDenyDial(addr.multiaddr, localCidrs)) {
-					skippedTransitive++;
-					continue;
-				}
-				const full = base.includes('/p2p/') ? base : `${base}/p2p/${pid}`;
-				collected.add(full);
-				perPeer++;
-			}
-		}
-		if (skippedSelf > 0 || skippedTransitive > 0) {
-			trace(`[NET] peer-announce filter: skipped ${skippedSelf} self + ${skippedTransitive} transitive non-routable addrs`);
-		}
-		if (collected.size === 0) return;
-		const lishTopics = this.pubsub.getTopics().filter((t: string) => t.startsWith(LISH_TOPIC_PREFIX));
-		if (lishTopics.length === 0) return;
-		const msg: PeerAnnounceMessage = { type: 'peer-announce', multiaddrs: Array.from(collected) };
-		trace(`[NET] peer-announce emit: ${collected.size} addrs (self + ${allPeers.length} known peers)`);
-		for (const topic of lishTopics) {
-			try {
-				await this.broadcast(topic, msg as unknown as Record<string, any>);
-			} catch (err: any) {
-				trace(`[NET] peer-announce publish failed topic=${topic}: ${err?.message ?? err}`);
-			}
-		}
-	}
-
-	/**
-	 * Accept inbound peer-announce: dial each advertised multiaddr through addBootstrapPeers
-	 * (which dedupes against our existing bootstrap set and skips our own peer ID).
-	 */
-	private async handlePeerAnnounce(data: PeerAnnounceMessage, networkID: string, fromPeerID?: string): Promise<void> {
-		if (!Array.isArray(data.multiaddrs) || data.multiaddrs.length === 0) return;
-		// Two-stage filter: shape (non-empty string) THEN routability (drop
-		// loopback + non-local private through shouldDenyDial). Without the
-		// routability stage, broadcasters with buggy emitters can inject their
-		// /ip4/127.0.0.1 into our bootstrap set, causing TCP loop → Noise
-		// identity-mismatch storm. Even though emitPeerAnnounce now filters
-		// these out on our side, we cannot trust older peers in the fleet to
-		// do the same — every receiver must be defensive.
-		const localCidrs = getLocalCidrs();
-		const rawCount = data.multiaddrs.length;
-		const filtered: string[] = [];
-		let droppedNonRoutable = 0;
-		for (const a of data.multiaddrs) {
-			if (typeof a !== 'string' || a.length === 0) continue;
-			if (filtered.length >= PEER_ANNOUNCE_MAX_TOTAL_ADDRS) break;
-			try {
-				if (shouldDenyDial(Multiaddr(a), localCidrs)) {
-					droppedNonRoutable++;
-					continue;
-				}
-			} catch {
-				// Unparseable multiaddr → drop (can't safely dial it anyway).
-				droppedNonRoutable++;
-				continue;
-			}
-			filtered.push(a);
-		}
-		if (filtered.length === 0) {
-			if (droppedNonRoutable > 0) trace(`[NET] peer-announce from ${fromPeerID?.slice(0, 16) ?? 'unknown'}: dropped all ${droppedNonRoutable}/${rawCount} addrs as non-routable`);
-			return;
-		}
-		trace(`[NET] peer-announce from ${fromPeerID?.slice(0, 16) ?? 'unknown'}: ${filtered.length}/${rawCount} addrs (dropped ${droppedNonRoutable} non-routable, network ${networkID.slice(0, 8)})`);
-		// Pass networkID so per-peer outcomes from gossiped entries surface in the
-		// UI under the network through which they arrived. Identity-mismatch
-		// outcomes inside addBootstrapPeers also trigger purgeStalePeer.
-		await this.addBootstrapPeers(filtered, networkID, 'discovered');
-		// Stamp `keep-alive-fleet` on every peer the announce mentioned. libp2p
-		// ReconnectQueue only acts on peers carrying a tag whose key starts with
-		// `keep-alive`; without this tag, fleet-discovered peers that drop are
-		// not re-dialed automatically. addBootstrapPeers() above tags via KEEP_ALIVE
-		// only for peer IDs it successfully extracts from multiaddrs — this adds
-		// the same treatment for every known peer, driving mesh maintenance.
-		if (this.node) {
-			for (const ma of filtered) {
-				try {
-					const mapath = Multiaddr(ma);
-					const pidComp = mapath.getComponents().find(c => c.code === 421);
-					const pid = pidComp?.value;
-					if (!pid) continue;
-					if (pid === this.node.peerId.toString()) continue;
-					await this.node.peerStore.merge(peerIDFromString(pid), {
-						tags: { 'keep-alive-fleet': { value: 50 } },
-					});
-				} catch {
-					/* invalid multiaddr — skip */
-				}
-			}
-		}
 	}
 
 	private setupBootstrapWorkaround(): void {
@@ -1289,7 +1090,7 @@ export class Network {
 					trace(`[NET] handleWant failed: ${err?.message ?? err}`);
 				});
 			} else if (data['type'] === 'peer-announce') {
-				this.handlePeerAnnounce(data as unknown as PeerAnnounceMessage, networkID, from).catch(err => {
+				this.peerAnnounce.handle(data as unknown as PeerAnnounceMessage, networkID, from).catch(err => {
 					trace(`[NET] handlePeerAnnounce failed: ${err?.message ?? err}`);
 				});
 			} else if (data['type'] === 'searchLishs') {
@@ -1559,10 +1360,7 @@ export class Network {
 			clearInterval(this.statusInterval);
 			this.statusInterval = null;
 		}
-		if (this.peerAnnounceInterval) {
-			clearTimeout(this.peerAnnounceInterval);
-			this.peerAnnounceInterval = null;
-		}
+		this.peerAnnounce.stop();
 		if (this.wantResponseCleanupInterval) {
 			clearInterval(this.wantResponseCleanupInterval);
 			this.wantResponseCleanupInterval = null;
